clean tests
