+++ pallets/subtensor/src/macros/hooks.rs
@@ -59,7 +59,7 @@ mod hooks {
                 .saturating_add(migrations::migrate_delete_subnet_3::migrate_delete_subnet_3::<T>())
                 // Doesn't check storage version. TODO: Remove after upgrade
                 // Storage version v5 -> v6
-                .saturating_add(migrations::migrate_total_issuance::migration5_total_issuance::<T>(false))
+                .saturating_add(migrations::migrate_total_issuance::migrate_total_issuance::<T>(false))
                 // Populate OwnedHotkeys map for coldkey swap. Doesn't update storage vesion.
                 // Storage version v6 -> v7
                 .saturating_add(migrations::migrate_populate_owned_hotkeys::migrate_populate_owned::<T>())
