+++ pallets/subtensor/src/staking.rs
@@ -44,10 +44,6 @@ impl<T: Config> Pallet<T> {
     ) -> dispatch::DispatchResult {
         // --- 1. We check the coldkey signuture.
         let coldkey = ensure_signed(origin)?;
-        ensure!(
-            !Self::coldkey_in_arbitration(&coldkey),
-            Error::<T>::ColdkeyIsInArbitration
-        );
         log::info!(
             "do_become_delegate( origin:{:?} hotkey:{:?}, take:{:?} )",
             coldkey,
@@ -137,10 +133,6 @@ impl<T: Config> Pallet<T> {
             hotkey,
             take
         );
-        ensure!(
-            !Self::coldkey_in_arbitration(&coldkey),
-            Error::<T>::ColdkeyIsInArbitration
-        );
 
         // --- 2. Ensure we are delegating a known key.
         //        Ensure that the coldkey is the owner.
@@ -212,10 +204,6 @@ impl<T: Config> Pallet<T> {
             coldkey,
             hotkey,
             take
-        );
-        ensure!(
-            !Self::coldkey_in_arbitration(&coldkey),
-            Error::<T>::ColdkeyIsInArbitration
         );
 
         // --- 2. Ensure we are delegating a known key.
@@ -301,10 +289,6 @@ impl<T: Config> Pallet<T> {
             coldkey,
             hotkey,
             stake_to_be_added
-        );
-        ensure!(
-            !Self::coldkey_in_arbitration(&coldkey),
-            Error::<T>::ColdkeyIsInArbitration
         );
 
         // Ensure the callers coldkey has enough stake to perform the transaction.
@@ -417,10 +401,6 @@ impl<T: Config> Pallet<T> {
             coldkey,
             hotkey,
             stake_to_be_removed
-        );
-        ensure!(
-            !Self::coldkey_in_arbitration(&coldkey),
-            Error::<T>::ColdkeyIsInArbitration
         );
 
         // Ensure that the hotkey account exists this is only possible through registration.
