+++ pallets/subtensor/src/swap/mod.rs
@@ -1,0 +1,3 @@
+use super::*;
+pub mod swap_coldkey;
+pub mod swap_hotkey;
