+++ pallets/subtensor/tests/root.rs
@@ -4,7 +4,7 @@ use crate::mock::*;
 use frame_support::{assert_err, assert_ok};
 use frame_system::Config;
 use frame_system::{EventRecord, Phase};
-use pallet_subtensor::migration;
+use pallet_subtensor::migrations;
 use pallet_subtensor::Error;
 use sp_core::{Get, H256, U256};
 
@@ -22,7 +22,7 @@ fn record(event: RuntimeEvent) -> EventRecord<RuntimeEvent, H256> {
 #[test]
 fn test_root_register_network_exist() {
     new_test_ext(1).execute_with(|| {
-        migration::migrate_create_root_network::<Test>();
+        migrations::migrate_create_root_network::migrate_create_root_network::<Test>();
         let hotkey_account_id: U256 = U256::from(1);
         let coldkey_account_id = U256::from(667);
         assert_ok!(SubtensorModule::root_register(
@@ -63,7 +63,7 @@ fn test_set_weights_not_root_error() {
 #[test]
 fn test_root_register_normal_on_root_fails() {
     new_test_ext(1).execute_with(|| {
-        migration::migrate_create_root_network::<Test>();
+        migrations::migrate_create_root_network::migrate_create_root_network::<Test>();
         // Test fails because normal registrations are not allowed
         // on the root network.
         let root_netuid: u16 = 0;
@@ -107,7 +107,7 @@ fn test_root_register_normal_on_root_fails() {
 #[test]
 fn test_root_register_stake_based_pruning_works() {
     new_test_ext(1).execute_with(|| {
-        migration::migrate_create_root_network::<Test>();
+        migrations::migrate_create_root_network::migrate_create_root_network::<Test>();
         // Add two networks.
         let root_netuid: u16 = 0;
         let other_netuid: u16 = 1;
@@ -196,7 +196,7 @@ fn test_root_register_stake_based_pruning_works() {
 fn test_root_set_weights() {
     new_test_ext(1).execute_with(|| {
         System::set_block_number(0);
-        migration::migrate_create_root_network::<Test>();
+        migrations::migrate_create_root_network::migrate_create_root_network::<Test>();
 
         let n: usize = 10;
         let root_netuid: u16 = 0;
@@ -338,7 +338,7 @@ fn test_root_set_weights() {
 fn test_root_set_weights_out_of_order_netuids() {
     new_test_ext(1).execute_with(|| {
         System::set_block_number(0);
-        migration::migrate_create_root_network::<Test>();
+        migrations::migrate_create_root_network::migrate_create_root_network::<Test>();
 
         let n: usize = 10;
         let root_netuid: u16 = 0;
@@ -458,7 +458,7 @@ fn test_root_set_weights_out_of_order_netuids() {
 fn test_root_subnet_creation_deletion() {
     new_test_ext(1).execute_with(|| {
         System::set_block_number(0);
-        migration::migrate_create_root_network::<Test>();
+        migrations::migrate_create_root_network::migrate_create_root_network::<Test>();
         // Owner of subnets.
         let owner: U256 = U256::from(0);
 
@@ -538,7 +538,7 @@ fn test_root_subnet_creation_deletion() {
 fn test_network_pruning() {
     new_test_ext(1).execute_with(|| {
         System::set_block_number(0);
-        migration::migrate_create_root_network::<Test>();
+        migrations::migrate_create_root_network::migrate_create_root_network::<Test>();
 
         assert_eq!(SubtensorModule::get_total_issuance(), 0);
 
@@ -630,7 +630,7 @@ fn test_network_pruning() {
 #[test]
 fn test_network_prune_results() {
     new_test_ext(1).execute_with(|| {
-        migration::migrate_create_root_network::<Test>();
+        migrations::migrate_create_root_network::migrate_create_root_network::<Test>();
 
         SubtensorModule::set_network_immunity_period(3);
         SubtensorModule::set_network_min_lock(0);
@@ -671,7 +671,7 @@ fn test_network_prune_results() {
 #[test]
 fn test_weights_after_network_pruning() {
     new_test_ext(1).execute_with(|| {
-        migration::migrate_create_root_network::<Test>();
+        migrations::migrate_create_root_network::migrate_create_root_network::<Test>();
 
         assert_eq!(SubtensorModule::get_total_issuance(), 0);
