+++ pallets/subtensor/tests/senate.rs
@@ -15,7 +15,7 @@ use sp_runtime::{
 use frame_system::pallet_prelude::*;
 use frame_system::Config;
 use pallet_collective::Event as CollectiveEvent;
-use pallet_subtensor::migration;
+use pallet_subtensor::migrations;
 use pallet_subtensor::Error;
 
 pub fn new_test_ext() -> sp_io::TestExternalities {
@@ -57,7 +57,7 @@ fn record(event: RuntimeEvent) -> EventRecord<RuntimeEvent, H256> {
 #[test]
 fn test_senate_join_works() {
     new_test_ext().execute_with(|| {
-        migration::migrate_create_root_network::<Test>();
+        migrations::migrate_create_root_network::migrate_create_root_network::<Test>();
 
         let netuid: u16 = 1;
         let tempo: u16 = 13;
@@ -125,7 +125,7 @@ fn test_senate_join_works() {
 #[test]
 fn test_senate_vote_works() {
     new_test_ext().execute_with(|| {
-        migration::migrate_create_root_network::<Test>();
+        migrations::migrate_create_root_network::migrate_create_root_network::<Test>();
 
         let netuid: u16 = 1;
         let tempo: u16 = 13;
@@ -233,7 +233,7 @@ fn test_senate_vote_works() {
 #[test]
 fn test_senate_vote_not_member() {
     new_test_ext().execute_with(|| {
-        migration::migrate_create_root_network::<Test>();
+        migrations::migrate_create_root_network::migrate_create_root_network::<Test>();
 
         let netuid: u16 = 1;
         let tempo: u16 = 13;
@@ -294,7 +294,7 @@ fn test_senate_vote_not_member() {
 #[test]
 fn test_senate_leave_works() {
     new_test_ext().execute_with(|| {
-        migration::migrate_create_root_network::<Test>();
+        migrations::migrate_create_root_network::migrate_create_root_network::<Test>();
 
         let netuid: u16 = 1;
         let tempo: u16 = 13;
@@ -362,7 +362,7 @@ fn test_senate_leave_works() {
 #[test]
 fn test_senate_leave_vote_removal() {
     new_test_ext().execute_with(|| {
-        migration::migrate_create_root_network::<Test>();
+        migrations::migrate_create_root_network::migrate_create_root_network::<Test>();
 
         let netuid: u16 = 1;
         let tempo: u16 = 13;
@@ -501,7 +501,7 @@ fn test_senate_leave_vote_removal() {
 #[test]
 fn test_senate_not_leave_when_stake_removed() {
     new_test_ext().execute_with(|| {
-        migration::migrate_create_root_network::<Test>();
+        migrations::migrate_create_root_network::migrate_create_root_network::<Test>();
 
         let netuid: u16 = 1;
         let tempo: u16 = 13;
@@ -582,7 +582,7 @@ fn test_senate_not_leave_when_stake_removed() {
 fn test_senate_join_current_delegate() {
     // Test that a current delegate can join the senate
     new_test_ext().execute_with(|| {
-        migration::migrate_create_root_network::<Test>();
+        migrations::migrate_create_root_network::migrate_create_root_network::<Test>();
 
         let netuid: u16 = 1;
         let tempo: u16 = 13;
@@ -656,7 +656,7 @@ fn test_senate_join_current_delegate() {
 fn test_adjust_senate_events() {
     // Test the events emitted after adjusting the senate successfully
     new_test_ext().execute_with(|| {
-        migration::migrate_create_root_network::<Test>();
+        migrations::migrate_create_root_network::migrate_create_root_network::<Test>();
 
         let netuid: u16 = 1;
         let tempo: u16 = 13;
